Move packages out of `cmd/`
This PR only moves packages out of `cmd/` following usual Go package
setup for a monorepo.
