+++ pkg/git/doc.go
