+++ pkg/httpclient/client_test.go
@@ -10,7 +10,7 @@ import (
 
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
-	"github.com/stacklok/toolhive-registry-server/cmd/thv-registry-api/pkg/httpclient"
+	"github.com/stacklok/toolhive-registry-server/pkg/httpclient"
 )
 
 func TestHTTPClient(t *testing.T) {
