+++ pkg/httpclient/types.go
