+++ pkg/sources/api_test.go
@@ -9,7 +9,7 @@ import (
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 
-	"github.com/stacklok/toolhive-registry-server/cmd/thv-registry-api/pkg/sources"
+	"github.com/stacklok/toolhive-registry-server/pkg/sources"
 	mcpv1alpha1 "github.com/stacklok/toolhive/cmd/thv-operator/api/v1alpha1"
 )
