+++ pkg/sources/git.go
@@ -9,7 +9,7 @@ import (
 
 	"sigs.k8s.io/controller-runtime/pkg/log"
 
-	"github.com/stacklok/toolhive-registry-server/cmd/thv-registry-api/pkg/git"
+	"github.com/stacklok/toolhive-registry-server/pkg/git"
 	mcpv1alpha1 "github.com/stacklok/toolhive/cmd/thv-operator/api/v1alpha1"
 )
