+++ pkg/sources/git_test.go
@@ -12,7 +12,7 @@ import (
 	"github.com/stretchr/testify/require"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
-	"github.com/stacklok/toolhive-registry-server/cmd/thv-registry-api/pkg/git"
+	"github.com/stacklok/toolhive-registry-server/pkg/git"
 	mcpv1alpha1 "github.com/stacklok/toolhive/cmd/thv-operator/api/v1alpha1"
 	"github.com/stacklok/toolhive/pkg/registry"
 )
